Fixes for Node16 moduleResolution
Node16 moduleResolution is stricter, and there is a bug in typeRoots
module resolution, so need to use `paths` config to resolve custom types
in @digidem/types.

Because of stricter resolution, Typescript no longer likes merging types
from both a `.js` and `.d.ts`. Switched `types.js` and `types.d.ts` to
be a single `.ts` file that defines the types. Using `.d.ts` for this
causes issues when compiling declarations for module packaging.