fix: Fix types and typescript config for publishing (#162)
* feat: Fix typescript errors

* Fixes for Node16 moduleResolution

Node16 moduleResolution is stricter, and there is a bug in typeRoots
module resolution, so need to use `paths` config to resolve custom types
in @digidem/types.

Because of stricter resolution, Typescript no longer likes merging types
from both a `.js` and `.d.ts`. Switched `types.js` and `types.d.ts` to
be a single `.ts` file that defines the types. Using `.d.ts` for this
causes issues when compiling declarations for module packaging.

* Configure TS declaration compilation

* configure npm export and files to pack

Using package.files instead of `.npmignore`

* Remove outdated example and old mapeo-schema dep

* fix drizzle table types

Needed to update drizzle for simplification of typescript docs

* fix: Manually written declaration for DataType

Typescript was not correctly compiling the declaration,
so switched to write it manually.

* fix: Add tests for exported types

Also includes e2e tests in `npm test`

* fix: commit drizzle migration SQL