feat: disable namespace replication
This enables the sync API for disabling sync by adding an "unreplicate"
function that removes a core from the replication stream, without
closing the stream. Unreplicating a core from a replication with a peer
is currently the only way to stop a peer from receiving data from you -
Hypercore has no API yet for blocking uploads to peers when they request
data.