fix: Fix sending RPC messages after connection (#354)
* fix: Fix sending RPC messages after connection

Trying to send an RPC message immediately after connecting the RPC
channel would fail because the protomux channel was not yet open. This
fix ensures that sending an RPC message waits for a peer to fully
connect before trying to send. This is necessary for sending device info
(e.g. the device name) immediately after connecting to a peer.

* fix type error

* fix invite-api for async rpc methods