feat: Add SyncController [2/2] (#348)
* cache state

* sum peer states in namespace state

* set cached state

* fix broken test

* feat: add PeerSyncController

Hypercore manages peers on a "per-hypercore" basis, e.g. each hypercore
has a unique set of peers. However for Mapeo we need to treat all
hypercores belonging to a peer as one, so we need to combine the peer
core sync states across all cores, and then have a means for a
particular peer to control which cores / namespaces are replicated to
the replication stream.

This abstraction is for managing the enabling of namespaces with a
particular peer, and will be managed by the sync controller. This
replaces functionality that is currently in the CoreManager and the
ReplicationStateMachine that is currently returned by
coreManager.replicate().

* cleanup and comment

* expose handleDiscoveryKey

* add tests

* update package-lock.json

* add sync control logic (tests broken)

* add downloading to peer controller

* pin hypercore version

* add sync controller

* add SyncController to MapeoProject

* WIP: fix up tests and start addressing bugs

* fix PeerSyncController tests

* fix bug

* fix types

* allow peer sync controller without peerId

* allow peer sync controller without peerId