feat: dataType.on('updated-docs') when docs update
Emits an array of updated docs for the data type whenever the docs
update (rather than emitting one doc at a time, it waits for indexing to
complete, then emits all the docs that have been updated. If there are
several updates to a particular docId, then only the resolved `head`
version is returned).

Attaching this listener to a data type that has many values / is
frequently updated will have a performance impact, because large amounts
of data could be read from the database and there is no limit, e.g. if
this was attached to the `observation` dataType and a new device synced
with a project with 10,000 observations, an attached listener would be
called with 10,000 observations.

The main reason for this feature is for internal use for capabilities,
listening for changes and modifying permissions accordingly.