feat: share deviceType with local peers
fixes #451 and #452. Does not store deviceType in project deviceInfo
records, which will be done in a follow up PR (see #455) which also
requires changes to @mapeo/schema. `deviceType` is optional, and must be
set in the constructor of MapeoManager. Ideally this should not be
possible to change, but not sure how to enforce that.