chore: fix flaky local peer device info test
`waitForPeers()` has a `waitForDeviceInfo` option. When enabled, this
helper should resolve only after device info is populated.

It had a tricky bug that only manifested some of the time (see
[example CI failure 1][1], [2]). In some cases, the function could
_think_ device info was populated when it wasn't.[^CI]

To fix this, I reworked the helper. It now does redundant work but it's
now stateless-ish, which should avoid this problem.

[^CI]: Well, that's what I could reproduce locally. The errors on CI are
       unhelpful ([which I hope to fix in Brittle][brittle PR]). It's
       possible a different problem exists on CI that I couldn't repro
       locally.

[1]: https://github.com/digidem/mapeo-core-next/actions/runs/7925562783/job/21731451900
[2]: https://github.com/digidem/mapeo-core-next/actions/runs/7965492933/job/21745121763
[brittle PR]: https://github.com/holepunchto/brittle/pull/44