chore: restore bitCount32 test (#489)
bitCount32's test is skipped because it is slow, and it's slow because
it iterates over all integers between 0 and 2^32.

This keeps that test skipped but adds a new test that we can run, which
only tests a few inputs. It also speeds up the test-only comparison
function by about 9x in my testing.