chore: update config import error
I had a problem where the config import tests failed. This seemed to be
because the `JSON.parse()` error message can change.

For example, `JSON.parse('{bad')` gives the following error message in
Firefox v123:

> expected property name or '}' at line 1 column 2 of the JSON data

But Node v20 gives the following:

> Expected property name or '}' in JSON at position 1

We now catch this error and return a consistent message, which should
make the tests more reliable. Because this code was getting a little
long, I moved it into its own function.