chore: save schema integers in INTEGER SQLite columns (#531)
If we use the ["integer" JSON Schema type][0], we should save that as an
INTEGER column, not a REAL column. Both will work for many numbers, but
larger integers (like `2**63`) can only be safely stored in INTEGERs.
This should be a bit more reliable.

As of this patch, we don't have any integer types in our schemas, so
this has no effect. However, that may change as we address [upcoming
issues][1].

[0]: https://json-schema.org/understanding-json-schema/reference/numeric#integer
[1]: https://github.com/digidem/mapeo-core-next/issues/150