chore: update Drizzle
This updates drizzle-orm and drizzle-kit to the latest versions.

Notably, [a drizzle-orm change][0] required a workaround to avoid
resolving a promise with something awaitable. `placeholder` was also
deprecated in favor of `sql.placeholder`.

I think this is a useful change on its own, but it also blocks upgrading
TypeScript to the latest version.

[0]: https://github.com/drizzle-team/drizzle-orm/commit/c063144dc08726cc15323582fe377210329e579e