chore: prefer strict Node assert
This change should have no impact on functionality.

`node:assert` has two modes: [strict] and [legacy]. The latter can have
some strange behavior. This adds an [ESLint rule][no-restricted-imports]
to prefer the strict mode, and fixes all violations.

I think this is a useful change on its own but it should make it a
little easier to drop Brittle in the future.

[strict]: https://nodejs.org/api/assert.html#strict-assertion-mode
[legacy]: https://nodejs.org/api/assert.html#legacy-assertion-mode
[no-restricted-imports]: https://eslint.org/docs/latest/rules/no-restricted-imports