test: make device info test more reliable
[When peers connect, they exchange device info.][0] If peers make
redundant connections[^1], they can send redundant device info messages.

[In tests, we connect all managers to each other, which includes
redundant connections.][2] As you might expect, this might exchange
redundant device info messages.

This test was failing because of the following sequence of events:

1. Manager A starts connecting to Manager B.
2. Manager B starts connecting to Manager A.
3. The first connection (from event 1) is established and Manager A
   sends its device info.
4. Manager B receives the device info.
5. In the test, `waitForPeers(managers, { waitForDeviceInfo: true })`
   resolves because Manager B has received device info.
6. The second connection (from event 2) is established and Manager A
   resends its device info.
7. In the test, Manager A changes its name and sends its device info
   again.
8. Manager B receives the second device info.
9. Manager B receives the final device info.

The penultimate step was the problem. Once `waitForPeers()` resolved,
our test expected that no more device info messages were in flight, but
that wasn't true.

The failure was intermittent but its failure became more likely when
testing on Node 20.[^2]

[^1]: This can happen if two peers discover each other at the same time
and both connect to each other. This is swiftly resolved, but not before
the redundant device info messages are sent.
[^2]: I don't know why.

[0]: https://github.com/digidem/mapeo-core-next/blob/2eda1269ee8287199c9d9d1ec0ad30a98fad100c/src/mapeo-manager.js#L269
[2]: https://github.com/digidem/mapeo-core-next/blob/2eda1269ee8287199c9d9d1ec0ad30a98fad100c/test-e2e/utils.js#L46-L53