fix: at startup, clean up left projects that failed to delete
If you start to leave a project and the operation fails partway, we now
finish the next time you try to load that project.

For example, imagine the following scenario:

1. The user presses the "leave project" button.
2. We mark them as having left the project, but haven't yet deleted any
   data.
3. Their phone runs out of battery.

Previously, the project state would be messed up in this scenario. Now,
we finish the cleanup before returning the project.

Fixes [#583].

[#583]: https://github.com/digidem/mapeo-core-next/issues/583