chore: use standard Namespace type
This is a types-only change.

The `Namespace` type is pretty central, but was stuffed away in
`core-manager/index.js`. Its type was also inferred in many cases.

Now it's in a more central spot: `types.ts`.

I think this is useful on its own, but it should make a future change
easier.