chore: move PRESYNC_NAMESPACES into central constants file
This change should have no impact.

The `PRESYNC_NAMESPACES` array is pretty central, but was stuffed away
in `sync/peer-sync-controller.js`; that doesn't quite feel like where it
belongs.

Now it's in a more central spot: `constants.js`.

I think this is useful on its own, but it should make a future change
easier.