chore: use separate helper file for NoiseSecretStreams (#816)
This change should have no impact on functionality.

This change creates `src/lib/noise-secret-stream-helpers.js`. It moves
`openedNoiseSecretStream()` and two types in there.

It also tweaks the types for `OpenedNoiseStream` and
`DestroyedNoiseStream`.

I think this is a useful change on its own but may also make an upcoming
change easier.