test: additional assertions in "blocked from sync" test (#864)
We test that blocked members shouldn't be able to sync. However, it's
possible that this test had false positives if the blocked member was
simply missing from the project entirely!

This adds two assertions to make sure that the other members see the
blocked one, and that they are indeed blocked.