chore: allow any JSON object to be saved in blob store
The blob store will need to store additional metadata, such as audio
duration, in the future. This changes its `metadata` type to any JSON
object to allow this.

It also adds a test for `blobStore.entry()` to ensure metadata is saved.

This change shouldn't affect the external API because the blob store is
not directly exposed. It's used internally by `BlobApi`, which did not
change.