chore: use more restrictive types for `timingSafeEqual`
This is a types-only change.

[`crypto.timingSafeEqual`][0] has some subtle behavior (highlighted in
bold red in the docs!) when passing a `Float32Array` or `Float64Array`.
Rather than adapt our wrapper to support this unusual situation, this
updates the types to avoid it.

We never took advantage of this unusual case, and now we are even less
likely to.

[0]: https://nodejs.org/api/crypto.html#cryptotimingsafeequala-b