feat: adding self-hosted servers (#952)
This adds support for adding self-hosted servers. It adds the following
functions:

- `project.$member.addServerPeer()`
- `project.$sync.connectServers()`
- `project.$sync.disconnectServers()`

This change doesn't include end-to-end tests for the server. That's
deliberate! We can't (easily) add those tests without the server being
released, but we can't (easily) release the server without this change.
Once this change is released, we can release the server, and then we
should be able to add end-to-end tests.

See [#886](https://github.com/digidem/comapeo-core/pull/886) for more.

Co-Authored-By: Gregor MacLennan <gmaclennan@digital-democracy.org>