feat: import legacy Mapeo data
This adds a new method to `MapeoProject`, `$importLegacyMapeoData`,
which takes a [legacy export format file][0] and imports its
observations into CoMapeo.

This will be used in [CoMapeo Headless][1] to help with importing.
Putting it here also makes it easy to add elsewhere in the future, such
as in the desktop app or the cloud server.

[0]: https://github.com/digidem/mapeo-legacy-export-format
[1]: https://github.com/digidem/comapeo-headless